Isolate @nestjs/testing dependency
