+++ src/app.provider.ts
@@ -2,7 +2,6 @@ import { INestApplication, VersioningType } from '@nestjs/common';
 import { DataSourceErrorFilter } from './routes/common/filters/data-source-error.filter';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { NestFactory } from '@nestjs/core';
-import { TestingModule } from '@nestjs/testing/testing-module';
 
 function configureVersioning(app: INestApplication) {
   app.enableVersioning({
@@ -10,7 +9,7 @@ function configureVersioning(app: INestApplication) {
   });
 }
 
-function configureShutdownHooks(app: INestApplication) {
+export function configureShutdownHooks(app: INestApplication) {
   app.enableShutdownHooks();
 }
 
@@ -28,7 +27,7 @@ function configureSwagger(app: INestApplication) {
   SwaggerModule.setup('', app, document);
 }
 
-const DEFAULT_CONFIGURATION: ((app: INestApplication) => void)[] = [
+export const DEFAULT_CONFIGURATION: ((app: INestApplication) => void)[] = [
   configureVersioning,
   configureShutdownHooks,
   configureFilters,
@@ -71,35 +70,5 @@ export class DefaultAppProvider extends AppProvider {
 
   protected getApp(module: any): Promise<INestApplication> {
     return NestFactory.create(module);
-  }
-}
-
-/**
- * A test {@link AppProvider}
- *
- * This provider provides an application given a {@link TestingModule}
- *
- * If the module provided is not a {@link TestingModule}, an error is thrown
- */
-export class TestAppProvider extends AppProvider {
-  // Disables shutdown hooks for tests (they are not required)
-  // Enabling this in the tests might result in a MaxListenersExceededWarning
-  // as the number of listeners that this adds exceed the default
-  protected readonly configuration: Array<(app: INestApplication) => void> =
-    DEFAULT_CONFIGURATION.filter((config) => config !== configureShutdownHooks);
-
-  constructor() {
-    super();
-    if (process.env.NODE_ENV !== 'test') {
-      throw Error('TestAppProvider used outside of a testing environment');
-    }
-  }
-
-  protected getApp(module: any): Promise<INestApplication> {
-    if (!(module instanceof TestingModule))
-      return Promise.reject(
-        `${module.constructor.name} is not a TestingModule`,
-      );
-    return Promise.resolve(module.createNestApplication());
   }
 }
