+++ src/routes/contracts/__tests__/get-contract.e2e-spec.ts
@@ -4,7 +4,7 @@ import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { AppModule } from '../../../app.module';
 import { redisClientFactory } from '../../../__tests__/redis-client.factory';
-import { TestAppProvider } from '../../../app.provider';
+import { TestAppProvider } from '../../../__tests__/test-app.provider';
 
 describe('Get contract e2e test', () => {
   let app: INestApplication;
