+++ src/routes/data-decode/__tests__/data-decode.e2e-spec.ts
@@ -4,7 +4,7 @@ import { Test } from '@nestjs/testing';
 import { RedisClientType } from 'redis';
 import * as request from 'supertest';
 import { AppModule } from '../../../app.module';
-import { TestAppProvider } from '../../../app.provider';
+import { TestAppProvider } from '../../../__tests__/test-app.provider';
 import { DataDecoded } from '../../../domain/data-decoder/entities/data-decoded.entity';
 import { redisClientFactory } from '../../../__tests__/redis-client.factory';
 import { getDataDecodedDtoBuilder } from '../entities/__tests__/get-data-decoded.dto.builder';
