+++ src/routes/safe-apps/__tests__/get-safe-apps.e2e-spec.ts
@@ -3,7 +3,7 @@ import { Test } from '@nestjs/testing';
 import { RedisClientType } from 'redis';
 import * as request from 'supertest';
 import { AppModule } from '../../../app.module';
-import { TestAppProvider } from '../../../app.provider';
+import { TestAppProvider } from '../../../__tests__/test-app.provider';
 import { redisClientFactory } from '../../../__tests__/redis-client.factory';
 
 describe('Get Safe Apps e2e test', () => {
